% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlights.R
\name{checkme}
\alias{checkme}
\title{Generate html code with highlighted string with line break}
\usage{
checkme(string, color = yellow)
}
\description{
Use this function with \code{params} function of Rmarkdown.
}
