% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlights.R
\name{highlight}
\alias{highlight}
\title{Generate html code with highlighted string}
\usage{
highlight(string, color = "skyblue")
}
\arguments{
\item{string}{String to wrap with html tag}

\item{color}{Color to use for background}
}
\description{
Use this function with \code{params} function of Rmarkdown.
}
